requirements.txt
Django==4.1.5
Pillow==9.4.0
PyJWT==2.6.0
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
boto3==1.26.59
botocore==1.29.59
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
defusedxml==0.7.1
distlib==0.3.6
django-allauth==0.52.0
django-appconf==1.0.5
django-author==1.2.0
django-classy-tags==4.0.0
django-easy-maps==1.1.3
django-frontend-forms==0.2.19
django-jquery-ui==1.11.4.1
django-jquery==3.1.0
django-location-field==2.1.0
django-material==1.11.3
django-mysql==4.8.0
django-phonenumber-field==7.0.2
django-ses==3.3.0
django-storages==1.13.2
django-timezone-field==5.0
filelock==3.9.0
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
icalendar==5.0.4
idna==3.4
jmespath==1.0.1
multidict==6.0.4
mysqlclient==2.1.1
oauthlib==3.2.2
openai==0.26.4
phonenumbers==8.13.5
platformdirs==2.6.2
pycodestyle==2.10.0
pycparser==2.21
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests-oauthlib==1.3.1
requests==2.28.2
s3transfer==0.6.0
setuptools-git==1.2
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
uWSGI==2.0.21
urllib3==1.26.14
virtualenv==20.17.1
yarl==1.8.2
